// utils/enums.ts
export type SignInEnum = {
  username: string;
  password: string;
};
export type SignUpEnum = {
  username: string;
  email: string;
  fullname: string;
  password: string;
};
export type SignUpAPIEnum = {
  username: string;
  email: string;
  firstname: string;
  lastname: string;
  password: string;
};